smartstore-price-web/requirements.txt
Django==5.2.4
pyyaml==6.0.2
gunicorn==23.0.0
requests==2.32.4
pybase64==1.4.1
bcrypt==4.3.0
selenium==4.34.0
beautifulsoup4==4.13.4